Print the max sum according to best index method
Best index method is a method which follows the following strategy
if a list is composed of 6 elements, then the best index method is an element +next 2 element+next 3 elements and so on till the list ends. For example,
A[0]+(A[1]+A[2])+(A[3]+A[4]+A[5])=sum1
A[1]+(A[2]+A[3])=sum 2 next three elements can't be included bcz the list index will be out of range.